requirements.txt
numpy>=1.21.0
matplotlib>=3.4.0
PyYAML>=5.4.1
stable-baselines3>=2.0.0
gymnasium>=0.28.1
torch>=1.13.0
tensorboard>=2.8.0
